+++ buchschloss/gui2/widgets.py
@@ -1,6 +1,5 @@
 """specific widgets and form helpers"""
 import functools
-import enum
 import operator
 import tkinter as tk
 import tkinter.ttk as ttk
@@ -364,138 +363,3 @@ class SearchResultWidget(mtk.ContainingWidget):
                 'wraplength': config.gui2.widget_size.main.width,
                 'command': partial(view_func, r)}))
         super().__init__(master, *widgets, direction=(tk.BOTTOM, tk.LEFT))
-
-
-class FormTag(enum.Enum):
-    SEARCH = '"search" action'
-    NEW = '"new" action'
-    EDIT = '"edit" action'
-    VIEW = '"view" action'
-
-
-class BaseForm(formlib.Form):
-    """Base class for forms, handling get_name, default content and autocompletes"""
-    form_name: str
-
-    def __init__(self, frame, tag, submit_callback):
-        super().__init__(frame, tag, submit_callback)
-        self.set_data(config.gui2.entry_defaults.get(self.form_name).mapping)
-
-    def __init_subclass__(cls, **kwargs):
-        """Handle autocompletes and set cls.form_name"""
-        cls.form_name = cls.__name__.replace('Form', '')
-        # This will put every widget spec into the standard form, required below
-        super().__init_subclass__(**kwargs)  # noqa -- it might accept kwargs later
-
-        for k, v in config.gui2.get('autocomplete').get(cls.form_name).mapping.items():
-            if k in cls.all_widgets:
-                for *_, w_kwargs in cls.all_widgets[k].values():
-                    w_kwargs.setdefault('autocomplete', v)
-
-    def get_name(self, name):
-        """redirect to utils.get_name inserting a form-specific prefix"""
-        return utils.get_name('::'.join(('form', self.form_name, self.tag.name, name)))
-
-
-class SearchForm(BaseForm):
-    """Add search options (and/or) + exact matching"""
-    all_widgets = {
-        'search_mode': {FormTag.SEARCH: (
-            formlib.RadioChoices, [(c, utils.get_name(c)) for c in ('and', 'or')], {})},
-        'exact_match': {FormTag.SEARCH: Checkbox},
-    }
-
-    def get_data(self):
-        """ignore empty data"""
-        if self.tag is FormTag.SEARCH:
-            return {k: v for k, v in super().get_data() if v or isinstance(v, bool)}
-        else:
-            return super().get_data()
-
-    def validate(self):
-        """ignore errors from empty widgets"""
-        errors = super().validate()
-        # NOTE: the password entry will raise ValueError if the passwords don't
-        # match, but it shouldn't be used in searches anyway.
-        # All other widgets shouldn't raise exceptions in .get()
-        if self.tag is FormTag.SEARCH:
-            data = self.get_data()
-            for k in errors.keys() - data.keys():
-                del errors[k]
-
-
-class AuthedForm(BaseForm):
-    """add a 'current_password' field for NEW and EDIT"""
-    all_widgets = {
-        'current_password': {
-            FormTag.NEW: PasswordEntry,
-            FormTag.EDIT: PasswordEntry,
-        }
-    }
-
-
-class EditForm(BaseForm):
-    """Adapt forms for the EDIT action.
-
-    On FormTag.EDIT:
-    Use OptionsFromSearch with setter=True for the first widget.
-    Modify ``.get_data`` to include the value of the first widget under ``'*args'``.
-    """
-    def __init_subclass__(cls, **kwargs):
-        super().__init_subclass__(**kwargs)
-        cls.id_name, widget_spec = next(iter(cls.all_widgets.items()))
-        if FormTag.EDIT in widget_spec:
-            raise TypeError("can't use EditForm if FormTag.EDIT is specified")
-        widget_spec[FormTag.EDIT] = (
-            OptionsFromSearch,
-            common.NSWithLogin(getattr(core, cls.form_name)),
-            {'default': None, 'setter': True},
-        )
-
-    def get_data(self):
-        """put the value of the ID widget under ``'*args'``"""
-        data = super().get_data()
-        if self.tag is FormTag.EDIT:
-            data['*args'] = (data.pop(self.id_name),)
-        return data
-
-
-class ViewForm(BaseForm):
-    """Adapt a form to be suitable with FormTag.VIEW
-
-    Don't show a submit button when used with FormTag.VIEW.
-
-    Insert display widgets (DisplayWidget or LinkWidget) on subclassing
-    where a specific widget for FormTag.VIEW is not specified.
-    The widget and arguments are chosen based on the default widget
-    or any other widget if the default is None:
-
-    - ``SearchMultiChoice`` creates a ``LinkWidget`` with ``multiple=True``
-    - ``OptionsFromSearch`` creates a normal ``LinkWidget``
-    - ``Checkbox`` creates a ``Checkbox`` with ``active=False``
-    - ``MultiChoicePopup`` creates a ``DisplayWidget`` with ``display='list'``
-    - everything else creates a ``DisplayWidget`` with ``display='str'``
-    """
-    def __init_subclass__(cls, **kwargs):
-        super().__init_subclass__(**kwargs)
-        for ws in cls.all_widgets.values():
-            w, *a, kw = next(
-                (w for w in (ws[None], *ws.values()) if w is not None),
-                (type(None), {}),
-            )
-            if issubclass(w, (SearchMultiChoice, OptionsFromSearch)):
-                ans = a[0] if a else kw.pop('action_ns')
-                new = (LinkWidget, ans, {'multiple': issubclass(w, SearchMultiChoice)})
-            elif issubclass(w, Checkbox):
-                new = (Checkbox, {'active': False})
-            else:
-                display = 'list' if issubclass(w, MultiChoicePopup) else 'str'
-                new = (DisplayWidget, display, {})
-            ws.setdefault(FormTag.VIEW, new)
-
-    def get_submit_widget(self):
-        """Don't show a submit button when used with FormTag.VIEW"""
-        if self.tag is FormTag.VIEW:
-            return None
-        else:
-            return super().get_submit_widget()
